+++ apps/commerce-api/src/test/kotlin/com/loopers/domain/order/OrderItemEntityTest.kt
@@ -1,6 +1,6 @@
 package com.loopers.domain.order
 
-import com.loopers.domain.order.OrderItemEntityFixture.Companion.anOrderItem
+import com.loopers.domain.order.fixture.OrderItemEntityFixture.Companion.anOrderItem
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Assertions.assertAll
 import org.junit.jupiter.api.DisplayName
