+++ apps/commerce-api/src/test/kotlin/com/loopers/domain/order/fixture/OrderEntityFixture.kt
@@ -1,11 +1,12 @@
-package com.loopers.domain.order
+package com.loopers.domain.order.fixture
 
+import com.loopers.domain.order.OrderEntity
 import com.loopers.domain.order.vo.OrderCustomer
-import com.loopers.domain.order.vo.OrderCustomerFixture.Companion.anOrderCustomer
+import com.loopers.domain.order.vo.OrderCustomerFixture
 
 class OrderEntityFixture {
     private var userId = 1L
-    private var orderCustomer = anOrderCustomer().build()
+    private var orderCustomer = OrderCustomerFixture.Companion.anOrderCustomer().build()
 
     companion object {
         fun anOrder(): OrderEntityFixture = OrderEntityFixture()
